fix: make nixvim use the same nixpkgs as home-manager
This caused some confusion when I tried to enable a package that's
unfree. By default nixvim builds its own instance of nixpkgs so we
either have to feed it our overlays/config or directly give it a pkgs.

I'm not sure which is best for stability. I opted for the solution with
less nixpkgs instances.